get researches в другой комопнент
